s3-sink/requirements.txt
quixstreams<2.10
python-dotenv
boto3